Prevent occasional exception when rendering logs
Sometimes when the user is scrolling the log entries, an NPE is thrown
and the log entries stop responding to clicks. That happens because the
Swing framework passes value=null in getTableCellRendererComponent and
we weren't dealing that case. I'm still not sure why that happens, but
now we try to get the log entry value from another source (the table
model) and if it also fails, we just ignore that row without throwing
an exception.

Signed-off-by: Crístian Deives <cristiandeives@gmail.com>