Fixes #31: Use system new line sep. in testPersistGroupMultipleFilters
The test FiltersRepositoryTests.testPersistGroupMultipleFilters splits
a string by '\n' and checks the result strings. On Windows, the new
line separator is '\r\n', so when the string is split by '\n', it
leaves a '\r' in each line, and the test fails due to that difference.

Split the string by using the new line separator of the current
system, instead of always using '\n', which isn't the separator
on Windows.