Underline filters that have a modified name
Underline the name of the filters which do not match the sarch pattern.
This is so it is easy to verify whether a search string for a filter is
what is actually displayed in the filters list.

Also show a tooltip displaying the actual search pattern so there is no
need to bring up the edit filter dialog just to know what the search
pattern is